Add makefile

+++ src/main.cpp
@@ -17,7 +17,7 @@ void save_mat(const char* output_file, const char* var, const Eigen::SparseMatri
 	// every column represents a pair of nodes
 	// every row represents a metapath
 	mat_sparse_t  sparse = { 0, };
-	size_t dims[2] = {rows, cols};
+	size_t dims[2] = {static_cast<size_t>(rows), static_cast<size_t>(cols)};
 	sparse.nzmax = (int)nnz;
 	sparse.nir = (int)nnz;
 	sparse.njc = (int)cols + 1;
